factor ShadowedKind out into renamer
