+++ crates/ruff_linter/src/renamer.rs
@@ -7,7 +7,10 @@ use ruff_diagnostics::Edit;
 use ruff_python_ast as ast;
 use ruff_python_codegen::Stylist;
 use ruff_python_semantic::{Binding, BindingKind, Scope, ScopeId, SemanticModel};
+use ruff_python_stdlib::{builtins::is_python_builtin, keyword::is_keyword};
 use ruff_text_size::Ranged;
+
+use crate::checkers::ast::Checker;
 
 pub(crate) struct Renamer;
 
@@ -368,4 +371,70 @@ impl Renamer {
             }
         }
     }
+}
+
+/// Enumeration of various ways in which a binding can shadow other variables
+#[derive(Debug, PartialEq, Eq, Copy, Clone)]
+pub(crate) enum ShadowedKind {
+    /// The variable shadows a global, nonlocal or local symbol
+    Some,
+    /// The variable shadows a builtin symbol
+    BuiltIn,
+    /// The variable shadows a keyword
+    Keyword,
+    /// The variable does not shadow any other symbols
+    None,
+}
+
+impl ShadowedKind {
+    /// Returns `true` if `self` shadows any global, nonlocal, or local symbol, keyword, or builtin.
+    pub(crate) fn shadows_any(&self) -> bool {
+        matches!(
+            self,
+            ShadowedKind::Some | ShadowedKind::BuiltIn | ShadowedKind::Keyword
+        )
+    }
+}
+
+/// Determines the kind of shadowing or conflict for a given variable name.
+///
+/// Returns `None` if `name` doesn't start with underscore `_`.
+///
+/// This function is useful for checking whether or not the `target` of a [`Rename::rename`] will
+/// shadow another binding.
+pub(crate) fn try_shadowed_kind(
+    name: &str,
+    checker: &Checker,
+    scope_id: ScopeId,
+) -> Option<ShadowedKind> {
+    // If the name doesn't start with an underscore, we don't consider it
+    if !name.starts_with('_') {
+        return None;
+    }
+
+    // Trim the leading underscores for further checks
+    let trimmed_name = name.trim_start_matches('_');
+
+    // Check the kind in order of precedence
+    if is_keyword(trimmed_name) {
+        return Some(ShadowedKind::Keyword);
+    }
+
+    if is_python_builtin(
+        trimmed_name,
+        checker.settings.target_version.minor(),
+        checker.source_type.is_ipynb(),
+    ) {
+        return Some(ShadowedKind::BuiltIn);
+    }
+
+    if !checker
+        .semantic()
+        .is_available_in_scope(trimmed_name, scope_id)
+    {
+        return Some(ShadowedKind::Some);
+    }
+
+    // Default to no shadowing
+    Some(ShadowedKind::None)
 }
